DEV-81 Moved TabLayout
