+++ src/layouts/TabLayout/TabEdge/TabEdge.module.scss
