+++ src/layouts/TabLayout/TabLayout.module.scss
@@ -30,3 +30,8 @@
   height: 100%;
   overflow: hidden;
 }
+
+.visibilityWrapper {
+  width: 100%;
+  height: 100%;
+}
